fix: move BooksForm to containers folder
